backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
python-multipart==0.0.17
pydantic==2.9.2
Pillow==10.4.0
pyzbar==0.1.9
